+++ src/CoreEx/Abstractions/IServiceCollectionExtensions.cs
@@ -164,26 +164,26 @@ namespace Microsoft.Extensions.DependencyInjection
         public static IServiceCollection AddEventPublisher<TEventPublisher>(this IServiceCollection services) where TEventPublisher : class, IEventPublisher => CheckServices(services).AddScoped<IEventPublisher, TEventPublisher>();
 
         /// <summary>
-        /// Adds the <see cref="LoggerEventSender"/> as the <see cref="IEventSender"/> scoped service.
+        /// Adds the <see cref="LoggerEventSender"/> as the <see cref="IEventSender"/> singleton service.
         /// </summary>
         /// <param name="services">The <see cref="IServiceCollection"/>.</param>
         /// <returns>The <see cref="IServiceCollection"/>.</returns>
         public static IServiceCollection AddLoggerEventSender(this IServiceCollection services) => CheckServices(services).AddEventSender<LoggerEventSender>();
 
         /// <summary>
-        /// Adds the <see cref="NullEventSender"/> as the <see cref="IEventSender"/> scoped service.
+        /// Adds the <see cref="NullEventSender"/> as the <see cref="IEventSender"/> singleton service.
         /// </summary>
         /// <param name="services">The <see cref="IServiceCollection"/>.</param>
         /// <returns>The <see cref="IServiceCollection"/>.</returns>
         public static IServiceCollection AddNullEventSender(this IServiceCollection services) => CheckServices(services).AddEventSender<NullEventSender>();
 
         /// <summary>
-        /// Adds the <typeparamref name="TEventSender"/> as the <see cref="IEventSender"/> scoped service.
+        /// Adds the <typeparamref name="TEventSender"/> as the <see cref="IEventSender"/> singleton service.
         /// </summary>
         /// <typeparam name="TEventSender">The <see cref="IEventSender"/> <see cref="Type"/>.</typeparam>
         /// <param name="services">The <see cref="IServiceCollection"/>.</param>
         /// <returns>The <see cref="IServiceCollection"/>.</returns>
-        public static IServiceCollection AddEventSender<TEventSender>(this IServiceCollection services) where TEventSender : class, IEventSender => CheckServices(services).AddScoped<IEventSender, TEventSender>();
+        public static IServiceCollection AddEventSender<TEventSender>(this IServiceCollection services) where TEventSender : class, IEventSender => CheckServices(services).AddSingleton<IEventSender, TEventSender>();
 
         /// <summary>
         /// Adds the <see cref="EventSubscriberOrchestrator"/> as a singleton service.
@@ -238,12 +238,12 @@ namespace Microsoft.Extensions.DependencyInjection
         });
 
         /// <summary>
-        /// Adds the <see cref="CoreEx.Text.Json.CloudEventSerializer"/> as the <see cref="IEventSerializer"/> scoped service.
+        /// Adds the <see cref="CoreEx.Text.Json.CloudEventSerializer"/> as the <see cref="IEventSerializer"/> singleton service.
         /// </summary>
         /// <param name="services">The <see cref="IServiceCollection"/>.</param>
         /// <param name="configure">The action to enable the <see cref="CoreEx.Text.Json.CloudEventSerializer"/> to be further configured.</param>
         /// <returns>The <see cref="IServiceCollection"/>.</returns>
-        public static IServiceCollection AddCloudEventSerializer(this IServiceCollection services, Action<IServiceProvider, CoreEx.Text.Json.CloudEventSerializer>? configure = null) => CheckServices(services).AddScoped<IEventSerializer>(sp =>
+        public static IServiceCollection AddCloudEventSerializer(this IServiceCollection services, Action<IServiceProvider, CoreEx.Text.Json.CloudEventSerializer>? configure = null) => CheckServices(services).AddSingleton<IEventSerializer>(sp =>
         {
             var ces = new CoreEx.Text.Json.CloudEventSerializer(sp.GetService<EventDataFormatter>());
             configure?.Invoke(sp, ces);
@@ -251,12 +251,12 @@ namespace Microsoft.Extensions.DependencyInjection
         });
 
         /// <summary>
-        /// Adds the <see cref="CoreEx.Text.Json.EventDataSerializer"/> as the <see cref="IEventSerializer"/> scoped service.
+        /// Adds the <see cref="CoreEx.Text.Json.EventDataSerializer"/> as the <see cref="IEventSerializer"/> singleton service.
         /// </summary>
         /// <param name="configure">The action to enable the <see cref="CoreEx.Text.Json.EventDataSerializer"/> to be further configured.</param>
         /// <param name="services">The <see cref="IServiceCollection"/>.</param>
         /// <returns>The <see cref="IServiceCollection"/>.</returns>
-        public static IServiceCollection AddEventDataSerializer(this IServiceCollection services, Action<IServiceProvider, CoreEx.Text.Json.EventDataSerializer>? configure = null) => CheckServices(services).AddScoped<IEventSerializer>(sp =>
+        public static IServiceCollection AddEventDataSerializer(this IServiceCollection services, Action<IServiceProvider, CoreEx.Text.Json.EventDataSerializer>? configure = null) => CheckServices(services).AddSingleton<IEventSerializer>(sp =>
         {
             var eds = new CoreEx.Text.Json.EventDataSerializer(sp.GetService<IJsonSerializer>(), sp.GetService<EventDataFormatter>());
             configure?.Invoke(sp, eds);
@@ -264,12 +264,12 @@ namespace Microsoft.Extensions.DependencyInjection
         });
 
         /// <summary>
-        /// Adds the <see cref="EventDataFormatter"/> as the scoped service.
+        /// Adds the <see cref="EventDataFormatter"/> as the singleton service.
         /// </summary>
         /// <param name="services">The <see cref="IServiceCollection"/>.</param>
         /// <param name="formatter">The optional <see cref="EventDataFormatter"/>; will default where not specified.</param>
         /// <returns>The <see cref="IServiceCollection"/>.</returns>
-        public static IServiceCollection AddEventDataFormatter(this IServiceCollection services, EventDataFormatter? formatter = null) => CheckServices(services).AddScoped(_ => formatter ?? new EventDataFormatter());
+        public static IServiceCollection AddEventDataFormatter(this IServiceCollection services, EventDataFormatter? formatter = null) => CheckServices(services).AddSingleton(_ => formatter ?? new EventDataFormatter());
 
         /// <summary>
         /// Adds the <see cref="FileLockSynchronizer"/> as the <see cref="IServiceSynchronizer"/> singleton service.